+++ libs/spark/src/Unstable_TextField/Unstable_TextField.tsx
@@ -135,10 +135,6 @@ export interface Unstable_TextFieldProps
    */
   SelectProps?: Partial<Unstable_SelectProps>;
   /**
-   * If `true`, the input will indicate a success.
-   */
-  success?: boolean;
-  /**
    * The content of the `endAdornment` (an `InputAdornment`), usually an `Icon`, `IconButton`, or `string`.
    */
   trailingEl?: ReactNode;
@@ -182,7 +178,7 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
       minRows,
       select = false,
       SelectProps,
-      size = 'medium',
+      size,
       success,
       trailingEl,
       type,
@@ -225,8 +221,6 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
         name={name}
         maxRows={maxRows}
         minRows={minRows}
-        size={size}
-        success={success}
         trailingEl={trailingEl}
         type={type}
         value={value}
@@ -250,15 +244,11 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
         ref={(ref as unknown) as RefObject<HTMLDivElement>}
         required={required}
         size={size}
+        success={success}
         {...other}
       >
         {label ? (
-          <Unstable_FormLabel
-            htmlFor={id}
-            id={labelId}
-            size={size}
-            {...FormLabelProps}
-          >
+          <Unstable_FormLabel htmlFor={id} id={labelId} {...FormLabelProps}>
             {label}
           </Unstable_FormLabel>
         ) : null}
@@ -268,7 +258,6 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
             aria-describedby={helperTextId}
             id={id}
             labelId={labelId}
-            size={size}
             value={value}
             input={InputElement}
             {...SelectProps}
@@ -280,11 +269,7 @@ const Unstable_TextField = forwardRef<unknown, Unstable_TextFieldProps>(
         )}
 
         {helperText ? (
-          <Unstable_FormHelperText
-            id={helperTextId}
-            size={size}
-            {...FormHelperTextProps}
-          >
+          <Unstable_FormHelperText id={helperTextId} {...FormHelperTextProps}>
             {helperText}
           </Unstable_FormHelperText>
         ) : null}