+++ libs/spark/CHANGELOG.md
@@ -32,6 +32,9 @@ API surface:
   - [feat] added "size" prop
     - values: `"small" | "medium"`
     - default: `"medium"`
+  - [feat] added "success" prop
+    - values: `true | false`
+    - default: `false`
 - **Unstable_FormHelperText**
   - [feat] added "size" prop
     - values: `"small" | "medium"`
@@ -71,6 +74,8 @@ API surface:
   - see **Unstable_FormControl**
 - **useFormControl**
   - [feat] added as a top-level directory
+- **useFormControl_unstable**
+  - [feat] init
 - **useRadioGroupMore**
   - [feat] added as a top-level directory
   - [breaking] removed (renamed to **useRadioGroupExtra_unstable**)